requirements.txt
python-dotenv==1.0.0
openai==0.28.1
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
pytest==8.0.0
pytest-asyncio==0.23.5 
aiofiles==23.2.1